+ list page is the default /puzzles route
